requirements.txt
# Core data science and ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
catboost>=1.1.0
scipy>=1.8.0

# Data processing and validation
pyyaml>=6.0
requests>=2.28.0
urllib3>=1.26.0
psutil>=5.9.0

# Web frameworks
flask>=2.2.0
flask-cors>=4.0.0
streamlit>=1.12.0

# MLOps and experiment tracking
mlflow>=2.0.0
dvc>=2.30.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0
mypy>=0.991

# Logging and utilities
python-dotenv>=0.19.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Performance testing
locust>=2.10.0

# Development utilities
jupyter>=1.0.0
ipykernel>=6.15.0

# Production deployment
gunicorn>=20.1.0
docker>=6.0.0

# Optional: Database connections
# sqlalchemy>=1.4.0
# psycopg2-binary>=2.9.0

# Optional: Cloud storage
# boto3>=1.24.0
# google-cloud-storage>=2.5.0